+++ src/main/java/wooteco/subway/admin/controller/page/IndexController.java
@@ -1,0 +1,13 @@
+package wooteco.subway.admin.controller.page;
+
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class IndexController {
+    @GetMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
+    public String index() {
+        return "service/index";
+    }
+}
