+++ src/main/resources/templates/service/map.html
@@ -41,7 +41,7 @@
         </div>
       </div>
     </div>
-    <script type="module" src="./js/App.js"></script>
-    <script type="module" src="./js/views/Edge.js"></script>
+    <script type="module" src="/service/js/App.js"></script>
+    <script type="module" src="/service/js/views/Map.js"></script>
   </body>
 </html>
